+++ NEWS.md
@@ -1,3 +1,13 @@
+## Features
+
+- Remove redefinition of `fillColor` and `drawColor` when symbolic color names
+  are used.
+
+## Bug Fixes
+
+ - Do not create color file if `tikzSymbolicColors` is off
+ - Fix issue with translating color with representation -1
+
 ## Bug fixes
 
 - Ignore fill color for lines to remove thin line (1 pixel wide) that was shown